student_ilya_zhuk/src/main/java/org/javaguru/travel/insurance/core/TravelCalculatePremiumServiceImpl.java
package org.javaguru.travel.insurance.core;

import org.javaguru.travel.insurance.rest.TravelCalculatePremiumRequest;
import org.javaguru.travel.insurance.rest.TravelCalculatePremiumResponse;
import org.springframework.stereotype.Component;
import java.time.LocalDate;

import java.math.BigDecimal;
import java.time.ZoneId;
import java.time.temporal.ChronoUnit;
import java.util.concurrent.TimeUnit;

@Component
public class TravelCalculatePremiumServiceImpl implements TravelCalculatePremiumService {

    @Override
    public TravelCalculatePremiumResponse calculatePremium(TravelCalculatePremiumRequest request) {
        TravelCalculatePremiumResponse response = new TravelCalculatePremiumResponse();
        response.setPersonFirstName(request.getPersonFirstName());
        response.setPersonLastName(request.getPersonLastName());
        response.setAgreementDateFrom(request.getAgreementDateFrom());
        response.setAgreementDateTo(request.getAgreementDateTo());
        response.setAgreementPrice(request.getAgreementPrice());
        long dist = request.getAgreementDateFrom().getTime() - request.getAgreementDateTo().getTime();
        var daysBetween = TimeUnit.DAYS.convert(dist, TimeUnit.MILLISECONDS);
        response.setAgreementPrice(new BigDecimal(daysBetween));


        return response;
    }
}
